rhi manual tests: Prevent warnings with MSVC
The QPair changes trigger warnings about size_t vs. quint32.

We made offsets and sizes 32-bit in the QRhi API to emphasize that
some of the graphics APIs are using 32-bit sizes still. It's a bit
unfortunate that pairs now generate warnings when the size does not
match. Just cast as needed.

+++ tests/manual/rhi/compressedtexture_bc1/compressedtexture_bc1.cpp
@@ -191,7 +191,7 @@ void Window::customRender()
     cb->setShaderResources();
     const QRhiCommandBuffer::VertexInput vbufBindings[] = {
         { d.vbuf, 0 },
-        { d.vbuf, 36 * 3 * sizeof(float) }
+        { d.vbuf, quint32(36 * 3 * sizeof(float)) }
     };
     cb->setVertexInput(0, 2, vbufBindings);
     cb->draw(36);
