+++ tests/manual/rhi/multiwindow_threaded/multiwindow_threaded.cpp
@@ -613,7 +613,7 @@ void Renderer::render(bool newlyExposed, bool wakeBeforePresent)
     cb->setShaderResources();
     const QRhiCommandBuffer::VertexInput vbufBindings[] = {
         { m_vbuf, 0 },
-        { m_vbuf, 36 * 3 * sizeof(float) }
+        { m_vbuf, quint32(36 * 3 * sizeof(float)) }
     };
     cb->setVertexInput(0, 2, vbufBindings);
     cb->draw(36);
